se agrega error en caso de no existir categoria

+++ src/main/java/com/alkemy/ong/services/impl/CategoriesServiceImpl.java
@@ -84,6 +84,9 @@ public class CategoriesServiceImpl implements CategoriesService {
     @Override
     public void deleted(String id) {
         Optional<Category>entity = this.categoryRepository.findById(id);
+        if(!entity.isEmpty()){
+            throw new RuntimeException("Category not present");
+        }
         this.categoryRepository.delete(entity.get());
     }
 }
